refactor: YoshiJava is the YoshiBase
