+++ src/releasers/java-lts.ts
@@ -20,11 +20,11 @@ import {VersionsMap} from '../updaters/update';
 
 // Java
 import {Version} from './java/version';
-import {JavaBase} from './java-base';
+import {JavaYoshi} from './java-yoshi';
 import {BranchName} from '../util/branch-name';
 import {PullRequestTitle} from '../util/pull-request-title';
 
-export class JavaLTS extends JavaBase {
+export class JavaLTS extends JavaYoshi {
   protected async coerceVersions(
     _cc: ConventionalCommits,
     _candidate: ReleaseCandidate,
@@ -56,50 +56,5 @@ export class JavaLTS extends JavaBase {
       previousTag: latestTag?.version,
       version,
     };
-  }
-
-  // Begin release configuration
-
-  // Override this method to use static branch names
-  // If you modify this, you must ensure that the releaser can parse the tag version
-  // from the pull request.
-  protected async buildBranchName(
-    _version: string,
-    includePackageName: boolean
-  ): Promise<BranchName> {
-    const defaultBranch = await this.gh.getDefaultBranch();
-    const packageName = await this.getPackageName();
-    if (includePackageName && packageName.getComponent()) {
-      return BranchName.ofComponentTargetBranch(
-        packageName.getComponent(),
-        defaultBranch
-      );
-    }
-    return BranchName.ofTargetBranch(defaultBranch);
-  }
-
-  // Override this method to modify the pull request title
-  protected async buildPullRequestTitle(
-    version: string,
-    includePackageName: boolean
-  ): Promise<string> {
-    const defaultBranch = await this.gh.getDefaultBranch();
-    const repoDefaultBranch = await this.gh.getRepositoryDefaultBranch();
-
-    // If we are proposing a release to a non-default branch, add the target
-    // branch in the pull request title.
-    // TODO: consider pushing this change up to the default pull request title
-    if (repoDefaultBranch === defaultBranch) {
-      return super.buildPullRequestTitle(version, includePackageName);
-    }
-    const packageName = await this.getPackageName();
-    const pullRequestTitle = includePackageName
-      ? PullRequestTitle.ofComponentTargetBranchVersion(
-          packageName.name,
-          defaultBranch,
-          version
-        )
-      : PullRequestTitle.ofTargetBranchVersion(defaultBranch, version);
-    return pullRequestTitle.toString();
   }
 }
